Split Settings and Footer into separate components
