+++ App/src/components/Settings/Settings.css
@@ -1,0 +1,34 @@
+.settings {
+  position: fixed;
+  top: 2rem;
+  right: 0;
+  left: 0;
+  margin: auto;
+  max-width: 42rem;
+  width: 50vw;
+  min-width: 20rem;
+  min-height: 40rem;
+  background: rgba(14, 26, 64, 0.98);
+  box-shadow: 1px 1px 4px black;
+  padding: 1rem;
+  z-index: 2;
+}
+
+.toggle-settings__button {
+  box-sizing: content-box;
+  display: block;
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  height: 2rem;
+  width: 2rem;
+  line-height: 1rem;
+  border-radius: 100%;
+  background: none;
+  color: white;
+  border: 1px solid white;
+  text-align: center;
+  cursor: pointer;
+  padding: 0.1rem;
+  font-size: 1rem;
+}
