+++ App/src/components/Testemony/testemony.js
@@ -3,15 +3,16 @@ import {
   deleteSession,
   getOrCreateSession,
   uploadTestemony,
-  generateConcatenatedVideo,
 } from "../../Services/vitneboksService";
 import "./testemony.css";
+import Settings from "../Settings/Settings";
 import {
   GetRecordingConstrains,
-  downoadFile,
+  downloadFile,
   getSrtFile,
   prepFile,
 } from "../../utilities";
+import Footer from "../Footer/Footer";
 
 const defaultQuestions = [
   "Hvordan føler du deg i dag etter dagens hendelser?",
@@ -50,16 +51,13 @@ const Testemony = () => {
   const [recordTime, setRecordTime] = useState(15000);
   const [waitTime, setWaitTime] = useState(30000);
   const [lastUpload, setLastUpload] = useState(null);
-  const [testemonialCount, setTestemonialCount] = useState(null);
+  const [testimonialCount, setTestimonialCount] = useState(null);
   const [actionShotCount, setActionShotCount] = useState(null);
 
   const [concatCompleted, setConcatCompleted] = useState(false);
   const [concatProcessStarted, setConcatProcessStarted] = useState(false);
   const [sessionWaiting, setSessionWaiting] = useState(false);
   const [sessionFetchTime, setSessionFetchTime] = useState(null);
-  const [sessionName, setSessionName] = useState(
-    localStorage.getItem("sessionName", "")
-  );
 
   const GetSession = useCallback(
     async (sessionKey = inputKey) => {
@@ -80,7 +78,7 @@ const Testemony = () => {
         setSessionKey(newSessionKey);
         setLastUpload(lastUpload);
         setActionShotCount(actionshots);
-        setTestemonialCount(testemonials);
+        setTestimonialCount(testemonials);
         setSharedKey(newSharedKey);
         setConcatCompleted(concatCompleted);
         localStorage.setItem("sessionKey", newSessionKey);
@@ -128,14 +126,6 @@ const Testemony = () => {
     }
   }, [recording, videoStream]);
 
-  const handleDownload = (url) => {
-    const link = document.createElement("a");
-    link.href = `${process.env.REACT_APP_API}${url}`;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
-
   const startRecording = async () => {
     setStarted(true);
     setCountdown(countdownTime / 1000);
@@ -178,8 +168,8 @@ const Testemony = () => {
 
           // Save video
           if (!sessionKey) {
-            downoadFile(videoBlob, videoFileName);
-            downoadFile(srtBlob, videoFileName.replace("mp4", "srt"));
+            downloadFile(videoBlob, videoFileName);
+            downloadFile(srtBlob, videoFileName.replace("mp4", "srt"));
           } else {
             // upload video
             await uploadTestemony(
@@ -257,30 +247,8 @@ const Testemony = () => {
       setSessionKey(null);
       setLastUpload(null);
       setActionShotCount(null);
-      setTestemonialCount(null);
+      setTestimonialCount(null);
     }
-  };
-
-  const handleTextareaChange = (event) => {
-    const inputString = event.target.value;
-    const newArray = parseNewlineSeparatedList(inputString);
-    if (newArray.length > 0) {
-      setQuestions(newArray);
-    } else {
-      setQuestions(defaultQuestions);
-    }
-
-    localStorage.setItem("questions", JSON.stringify(newArray));
-    setQuestionsRawString(inputString);
-    localStorage.setItem("questionsRawString", JSON.stringify(inputString));
-  };
-
-  const parseNewlineSeparatedList = (inputString) => {
-    const arrayOfStrings = inputString.split("\n");
-    const trimmedArray = arrayOfStrings
-      ?.map((str) => str.trim())
-      .filter((str) => str !== "");
-    return trimmedArray;
   };
 
   return (
@@ -428,332 +396,35 @@ const Testemony = () => {
       </div>
 
       {settingsOpen && (
-        <aside
-          className="settings"
-          style={{
-            position: "fixed",
-            top: "2rem",
-            right: "0",
-            left: "0",
-            margin: "auto",
-            maxWidth: "42rem",
-            width: "50vw",
-            minWidth: "20rem",
-            minHeight: "40rem",
-            background: "rgb(14 26 64 / 98%)",
-            boxShadow: "1px 1px 4px black",
-            padding: "1rem",
-            zIndex: 2,
-          }}
-        >
-          <button
-            style={{
-              boxSizing: "content-box",
-              display: "block",
-              position: "absolute",
-              top: "1rem",
-              right: "1rem",
-              height: "2rem",
-              width: "2rem",
-              lineHeight: "1rem",
-              borderRadius: "100%",
-              background: "none",
-              color: "white",
-              border: "1px solid white",
-              textAlign: "center",
-              cursor: "pointer",
-              padding: ".1rem",
-              fontSize: "1rem",
-            }}
-            aria-label="Vis bruksanvisning"
-            onClick={() => {
-              setSettingsOpen(false);
-            }}
-          >
-            X
-          </button>
-          <div className="form">
-            <h3>Konfigurasjon</h3>
-            <div>
-              <span>Ventetid før opptak:</span>
-              <input
-                min={1}
-                max={60}
-                type="number"
-                value={countdownTime / 1000}
-                onChange={(e) => {
-                  let value = parseInt(e.target.value, 10) * 1000;
-                  setCountdownTime(value);
-                  localStorage.setItem("countdownTime", value);
-                }}
-              />
-            </div>
-            <div>
-              <span>Opptakstid:</span>
-              <input
-                type="number"
-                value={recordTime / 1000}
-                max={30}
-                min={5}
-                onChange={(e) => {
-                  let value = parseInt(e.target.value, 10) * 1000;
-                  setRecordTime(value);
-                  localStorage.setItem("recordTime", value);
-                }}
-              />
-            </div>
-            <div>
-              <span>Ventetid etter opptak:</span>
-              <input
-                type="number"
-                value={waitTime / 1000}
-                min={1}
-                max={600}
-                onChange={(e) => {
-                  let value = parseInt(e.target.value, 10) * 1000;
-                  setWaitTime(value);
-                  localStorage.setItem("waitTime", value);
-                }}
-              />
-            </div>
-
-            <span>Spørsmål (ett per linje)</span>
-            <textarea
-              onChange={handleTextareaChange}
-              value={questionsRawString ?? ""}
-              style={{
-                width: "100%",
-                minHeight: "5rem",
-                color: "black",
-                textAlign: "left",
-              }}
-            />
-            {sessionKey == null ? (
-              <React.Fragment>
-                <div>
-                  <span>Opprett ny vitneboks:</span>
-                  <button onClick={() => GetSession(null)}>Opprett ny</button>
-                </div>
-                <div>
-                  <span>Har du allerede en vitneboks?</span>
-                  <div>
-                    <input
-                      type="text"
-                      style={{ width: "5rem" }}
-                      value={inputKey}
-                      placeholder="vitnboks-ID"
-                      onChange={(e) => setInputKey(e.target.value)}
-                    />
-                    &nbsp;
-                    <button onClick={() => GetSession()}>
-                      Koble til eksisterende
-                    </button>
-                  </div>
-                </div>
-              </React.Fragment>
-            ) : (
-              <React.Fragment>
-                <div>
-                  <h3>Tilkobling</h3>
-                  <span>
-                    Sist sjekket{" "}
-                    {new Date(sessionFetchTime).toLocaleTimeString()}{" "}
-                    <span
-                      className={`clickable ${sessionWaiting ? "spinner" : ""}`}
-                      onClick={() => GetSession()}
-                    >
-                      🔄️
-                    </span>
-                  </span>
-                </div>
-                <div>
-                  <span>Arrangementnavn:</span>
-                  <input
-                    type="test"
-                    value={sessionName}
-                    onChange={(e) => {
-                      setSessionName(e.target.value);
-                      localStorage.setItem("sessionName", e.target.value);
-                    }}
-                  />
-                </div>
-                <div>
-                  <span>Antall videoer:</span>
-                  <div>
-                    <div
-                      style={{
-                        display: "grid",
-                        gridTemplateColumns: "1fr 1fr",
-                      }}
-                    >
-                      <div>
-                        <span>Vitnesbyrd: {testemonialCount}</span>
-                        <br />
-                        <span>Delelink: {actionShotCount}</span>
-                        <br />
-                      </div>
-                      <button
-                        onClick={() =>
-                          handleDownload(
-                            `download-session-files?sessionKey=${sessionKey}`
-                          )
-                        }
-                      >
-                        Last ned alle filer
-                      </button>
-                    </div>
-                  </div>
-                </div>
-                {lastUpload && (
-                  <div>
-                    <span>Siste opplasting:</span>
-                    <span>{new Date(lastUpload).toLocaleString()}</span>
-                  </div>
-                )}
-                {testemonialCount + actionShotCount >= 1 &&
-                  !concatCompleted && (
-                    <div>
-                      <span>
-                        Vitneboksvideoen <br /> - Samle alle videoene til én
-                        fil.
-                      </span>
-                      <button
-                        className="button"
-                        disabled={concatProcessStarted}
-                        onClick={async () => {
-                          setConcatProcessStarted(true);
-                          await generateConcatenatedVideo(
-                            sessionKey,
-                            sessionName
-                          );
-                          setConcatProcessStarted(false);
-                          GetSession(sessionKey);
-                        }}
-                      >
-                        {!concatProcessStarted ? (
-                          "Lag video"
-                        ) : (
-                          <span className="spinner">🤖</span>
-                        )}
-                      </button>
-                    </div>
-                  )}
-                {actionShotCount + testemonialCount >= 1 && concatCompleted && (
-                  <div>
-                    <span>
-                      Vitneboksvideoen <br /> - Samle alle videoene til én fil.
-                    </span>
-                    <button
-                      onClick={() =>
-                        handleDownload(
-                          `download-concatenated-video?sessionKey=${sessionKey}`
-                        )
-                      }
-                    >
-                      Last ned
-                    </button>
-                  </div>
-                )}
-                <div>
-                  <span>Vitneboks-ID:</span>
-                  <input type="text" value={sessionKey} disabled={true} />
-                </div>
-                <div>
-                  <span>Delelink:</span>
-                  <input
-                    type="text"
-                    value={`${window.location}?session=${sharedKey}`}
-                    disabled={true}
-                  />
-                </div>
-                <div>
-                  <span></span>
-                  <button className="red" onClick={deleteSessionClick}>
-                    Slett vitneboks
-                  </button>
-                </div>
-              </React.Fragment>
-            )}
-          </div>
-        </aside>
+        <Settings
+          setQuestionsRawString={setQuestionsRawString}
+          defaultQuestions={defaultQuestions}
+          setSettingsOpen={setSettingsOpen}
+          countdownTime={countdownTime}
+          setCountdownTime={setCountdownTime}
+          waitTime={waitTime}
+          recordTime={recordTime}
+          setRecordTime={setRecordTime}
+          setWaitTime={setWaitTime}
+          questionsRawString={questionsRawString}
+          inputKey={inputKey}
+          sessionKey={sessionKey}
+          GetSession={GetSession}
+          setInputKey={setInputKey}
+          sessionFetchTime={sessionFetchTime}
+          sessionWaiting={sessionWaiting}
+          testimonialCount={testimonialCount}
+          actionShotCount={actionShotCount}
+          lastUpload={lastUpload}
+          setConcatProcessStarted={setConcatProcessStarted}
+          concatProcessStarted={concatProcessStarted}
+          concatCompleted={concatCompleted}
+          sharedKey={sharedKey}
+          deleteSessionClick={deleteSessionClick}
+          setQuestions={setQuestions}
+        />
       )}
-      <div
-        className="footer"
-        style={{
-          position: "fixed",
-          display: "flex",
-          alignItems: "baseline",
-          bottom: "0",
-          left: "5%",
-          right: "5%",
-        }}
-      >
-        <p
-          style={{
-            bottom: "0",
-            left: "5%",
-            alignItems: "baseline",
-            alignContent: "baseline",
-          }}
-        >
-          Sponset av
-          <a style={{ margin: "0 5px" }} href="https://spritjakt.no">
-            <img
-              src="spritjakt.png"
-              alt="spritjakt logo"
-              style={{ margin: "5px" }}
-              height={"25px"}
-            />
-          </a>
-          og
-          <a
-            style={{ margin: "0 5px" }}
-            href="https://erdetfesthosmatsikveld.no"
-          >
-            erdetfesthosmatsikveld.no
-          </a>
-        </p>
-        <p
-          style={{
-            right: "0%",
-            left: "0%",
-            bottom: "0",
-            marginLeft: "auto",
-            marginRight: "auto",
-            alignItems: "baseline",
-            alignContent: "baseline",
-            gap: "10px",
-          }}
-        >
-          © 2024{" "}
-          <a
-            target="_blank"
-            rel="noopener noreferrer"
-            href="https://no.linkedin.com/in/mats-l%C3%B8vstrand-berntsen-4682b2142"
-          >
-            Mats Løvstrand Berntsen
-          </a>
-        </p>
-        <p
-          style={{
-            display: "flex",
-            bottom: "0",
-            right: "5%",
-            alignItems: "baseline",
-            alignContent: "baseline",
-            gap: "10px",
-          }}
-        >
-          Kildekoden finner du på
-          <a
-            target="_blank"
-            rel="noopener noreferrer"
-            href="https://github.com/matslb/vitneboksen"
-          >
-            Github
-          </a>
-        </p>
-      </div>
+      <Footer />
     </div>
   );
 };
